# specs/001-pet-browsing/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: PetStore Pet Browsing API
  version: 1.0.0
  description: Read-only pet browsing endpoints. Sorting is intentionally not supported.
servers:
  - url: /tayag/api/v1
paths:
  /pets:
    get:
      summary: Browse available pets
      description: Returns available pets only. User-selectable sorting is not supported.
      parameters:
        - name: category
          in: query
          required: false
          description: Optional supported category filter.
          schema:
            $ref: "#/components/schemas/PetCategory"
      responses:
        "200":
          description: Available pets for all categories or one category.
          content:
            application/json:
              schema:
                type: object
                required:
                  - pets
                properties:
                  pets:
                    type: array
                    items:
                      $ref: "#/components/schemas/PetSummary"
        "400":
          description: Unsupported category value.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pets/{petId}:
    get:
      summary: View pet details
      description: Returns detail for an available pet or a clear unavailable state.
      parameters:
        - name: petId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Available pet detail.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PetDetail"
        "404":
          description: Pet was not found or is not available for browsing.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /pet-categories:
    get:
      summary: List supported pet categories
      responses:
        "200":
          description: Supported category values and display labels.
          content:
            application/json:
              schema:
                type: object
                required:
                  - categories
                properties:
                  categories:
                    type: array
                    items:
                      type: object
                      required:
                        - value
                        - label
                      properties:
                        value:
                          $ref: "#/components/schemas/PetCategory"
                        label:
                          type: string
components:
  schemas:
    PetCategory:
      type: string
      enum:
        - DOG
        - CAT
        - BIRD
        - REPTILE
        - FISH
    AvailabilityState:
      type: string
      enum:
        - AVAILABLE
        - RESERVED
        - SOLD
        - UNAVAILABLE
        - REMOVED
    PetSummary:
      type: object
      required:
        - id
        - name
        - category
        - price
        - availability
      properties:
        id:
          type: string
        name:
          type: string
        category:
          $ref: "#/components/schemas/PetCategory"
        price:
          type: number
          format: double
          minimum: 0
        availability:
          $ref: "#/components/schemas/AvailabilityState"
        summary:
          type: string
        primaryMediaUrl:
          type: string
          nullable: true
    PetDetail:
      allOf:
        - $ref: "#/components/schemas/PetSummary"
        - type: object
          properties:
            description:
              type: string
            careNotes:
              type: string
              nullable: true
            updatedAt:
              type: string
              format: date-time
    ErrorResponse:
      type: object
      required:
        - message
      properties:
        message:
          type: string